[GTNAP-618] wip: Adding HBAWS to Exemplo

+++ contrib/hbaws/src/hbaws.h
@@ -61,31 +61,31 @@ extern HB_BOOL hb_aws_s3_restore(HB_ITEM *bucket_block, HB_ITEM *key_block, cons
 
 extern int hb_aws_s3_size(const S3Objs *objs);
 
-extern const char *hb_aws_s3_key(const S3Objs *objs, int i);
+extern const char *hb_aws_s3_key(const S3Objs *objs, size_t i);
 
-extern long long hb_aws_s3_content_size(const S3Objs *objs, int i);
+extern long long hb_aws_s3_content_size(const S3Objs *objs, size_t i);
 
-extern const char *hb_aws_s3_content_type(const S3Objs *objs, int i);
+extern const char *hb_aws_s3_content_type(const S3Objs *objs, size_t i);
 
-extern const char *hb_aws_s3_date(const S3Objs *objs, int i);
+extern const char *hb_aws_s3_date(const S3Objs *objs, size_t i);
 
-extern const char *hb_aws_s3_time(const S3Objs *objs, int i);
+extern const char *hb_aws_s3_time(const S3Objs *objs, size_t i);
 
-extern const char *hb_aws_s3_timezone(const S3Objs *objs, int i);
+extern const char *hb_aws_s3_timezone(const S3Objs *objs, size_t i);
 
-extern const char *hb_aws_s3_storage_class(const S3Objs *objs, int i);
+extern const char *hb_aws_s3_storage_class(const S3Objs *objs, size_t i);
 
-extern HB_BOOL hb_aws_s3_is_restore(const S3Objs *objs, int i);
+extern HB_BOOL hb_aws_s3_is_restore(const S3Objs *objs, size_t i);
 
-extern const char *hb_aws_s3_restore_date(const S3Objs *objs, int i);
+extern const char *hb_aws_s3_restore_date(const S3Objs *objs, size_t i);
 
-extern const char *hb_aws_s3_restore_time(const S3Objs *objs, int i);
+extern const char *hb_aws_s3_restore_time(const S3Objs *objs, size_t i);
 
-extern const char *hb_aws_s3_restore_timezone(const S3Objs *objs, int i);
+extern const char *hb_aws_s3_restore_timezone(const S3Objs *objs, size_t i);
 
-extern const char *hb_aws_s3_checksum_algorithm(const S3Objs *objs, int i);
+extern const char *hb_aws_s3_checksum_algorithm(const S3Objs *objs, size_t i);
 
-extern const char *hb_aws_s3_etag(const S3Objs *objs, int i);
+extern const char *hb_aws_s3_etag(const S3Objs *objs, size_t i);
 
 HB_EXTERN_END
